Refactor: Correct the locations of library, sumnadasa, wala canteen, sentra, material, mechnaical, chemical, integrated departments
